semantic_memory_baseline/exercises/cossim_cluster/requirements.txt
torch
transformers
decord
numpy
pillow
torchvision
torchaudio
scikit-learn
einops
matplotlib
openai-clip
openai